improve Earth Tremor to be on par with tabletop
