+++ SolastaUnfinishedBusiness/Translations/de/Spells/Spells01-de.txt
@@ -20,6 +20,8 @@ Feedback/&AdditionalDamageWrathfulSmiteFormat=Zorniger Smite!
 Feedback/&AdditionalDamageWrathfulSmiteLine={0} fügt {1} durch einen zornigen Schlag (+{2}) mehr Schaden zu
 Spell/&ChromaticOrbDescription=Sie schleudern eine Energiekugel mit einem Durchmesser von 4 Zoll auf eine Kreatur, die Sie in Reichweite sehen können. Du wählst Säure, Kälte, Feuer, Blitz, Gift oder Donner für die Art der Kugel, die du erschaffst, und führst dann einen Fernzauberangriff gegen das Ziel aus. Wenn der Angriff trifft, erleidet die Kreatur 3W8 Schaden der gewählten Art.
 Spell/&ChromaticOrbTitle=Chromatische Kugel
+Spell/&EarthTremorDescription=Du triffst auf den Boden und entfesselst ein Beben seismischer Kraft, das Erde, Felsen und Sand aufwirbelt.
+Spell/&EarthTremorTitle=Erdbeben
 Spell/&EnsnaringStrikeDescription=Dein nächster Treffer verursacht zusätzlichen Stichschaden und dein Ziel wird von Ranken umgarnt.
 Spell/&EnsnaringStrikeTitle=Verstrickender Schlag
 Spell/&MuleDescription=Der Empfänger dieses Zaubers kann die Auswirkungen schwerer Lasten oder Rüstungen auf die Bewegungsgeschwindigkeit ignorieren. Sie können auch etwas mehr Gewicht tragen.
