+++ SolastaUnfinishedBusiness/Translations/de/Spells/Spells03-de.txt
@@ -12,8 +12,6 @@ Feedback/&AdditionalDamageSpiritShroudDamageRadiantFormat=Geisterschleier!
 Feedback/&AdditionalDamageSpiritShroudDamageRadiantLine={0} fügt {1} durch Spirit Shroud {2} strahlenden Schaden zu!
 Spell/&BlindingSmiteDescription=Ihr nächster Treffer verursacht zusätzlichen Strahlungsschaden, wobei Ihre Waffe mit einem hellen Licht aufblitzt. Außerdem muss dem Ziel ein Konstitutionsrettungswurf gelingen oder für 1 Minute geblendet werden.
 Spell/&BlindingSmiteTitle=Blendendes Smite
-Spell/&EarthTremorDescription=Du triffst auf den Boden und entfesselst ein Beben seismischer Kraft, das Erde, Felsen und Sand aufwirbelt.
-Spell/&EarthTremorTitle=Erdbeben
 Spell/&SpiritShroudDamageColdDescription=Füge zusätzlichen Kälteschaden zu.
 Spell/&SpiritShroudDamageColdTitle=Geisterschleier (kalt)
 Spell/&SpiritShroudDamageNecroticDescription=Füge zusätzlichen nekrotischen Schaden zu.
