+++ SolastaUnfinishedBusiness/Translations/en/Spells/Spells01-en.txt
@@ -20,6 +20,8 @@ Feedback/&AdditionalDamageWrathfulSmiteFormat=Wrathful Smite!
 Feedback/&AdditionalDamageWrathfulSmiteLine={0} deals more damage to {1} through a wrathful smite (+{2})
 Spell/&ChromaticOrbDescription=You hurl a 4-inch-diameter sphere of energy at a creature that you can see within range. You choose acid, cold, fire, lightning, poison, or thunder for the type of orb you create, and then make a ranged spell attack against the target. If the attack hits, the creature takes 3d8 damage of the type you chose.
 Spell/&ChromaticOrbTitle=Chromatic Orb
+Spell/&EarthTremorDescription=You strike the ground and unleash a tremor of seismic force, hurling up earth, rock, and sand.
+Spell/&EarthTremorTitle=Earth Tremor
 Spell/&EnsnaringStrikeDescription=Your next hit causes additional piercing damage and your target is ensnared by vines.
 Spell/&EnsnaringStrikeTitle=Ensnaring Strike
 Spell/&MuleDescription=The recipient of this spell is able to ignore the effects of heavy loads or armor on movement speed. They can also carry slightly more weight.
