+++ SolastaUnfinishedBusiness/Translations/en/Spells/Spells03-en.txt
@@ -12,8 +12,6 @@ Feedback/&AdditionalDamageSpiritShroudDamageRadiantFormat=Spirit Shroud!
 Feedback/&AdditionalDamageSpiritShroudDamageRadiantLine={0} deals {2} radiant damage to {1} from Spirit Shroud!
 Spell/&BlindingSmiteDescription=Your next hit causes additional radiant damage with your weapon flaring with a bright light. Additionally, the target must succeed on a Constitution saving throw or be blinded for 1 minute.
 Spell/&BlindingSmiteTitle=Blinding Smite
-Spell/&EarthTremorDescription=You strike the ground and unleash a tremor of seismic force, hurling up earth, rock, and sand.
-Spell/&EarthTremorTitle=Earth Tremor
 Spell/&SpiritShroudDamageColdDescription=Deal bonus cold damage.
 Spell/&SpiritShroudDamageColdTitle=Spirit Shroud (Cold)
 Spell/&SpiritShroudDamageNecroticDescription=Deal bonus necrotic damage.
