+++ SolastaUnfinishedBusiness/Translations/es/Spells/Spells01-es.txt
@@ -20,6 +20,8 @@ Feedback/&AdditionalDamageWrathfulSmiteFormat=¡Castigo colérico!
 Feedback/&AdditionalDamageWrathfulSmiteLine={0} inflige más daño a {1} a través de un golpe airado (+{2})
 Spell/&ChromaticOrbDescription=Lanzas una esfera de energía de 4 pulgadas de diámetro a una criatura que puedes ver dentro del alcance. Eliges ácido, frío, fuego, relámpago, veneno o trueno para el tipo de orbe que creas y luego realizas un ataque de hechizo a distancia contra el objetivo. Si el ataque impacta, la criatura sufre 3d8 puntos de daño del tipo que hayas elegido.
 Spell/&ChromaticOrbTitle=Orbe cromático
+Spell/&EarthTremorDescription=Golpeas el suelo y desatas un temblor de fuerza sísmica, arrojando tierra, rocas y arena.
+Spell/&EarthTremorTitle=Terremoto
 Spell/&EnsnaringStrikeDescription=Tu próximo golpe causa daño perforante adicional y tu objetivo queda atrapado en enredaderas.
 Spell/&EnsnaringStrikeTitle=Golpe atrapante
 Spell/&MuleDescription=El destinatario de este hechizo puede ignorar los efectos de cargas pesadas o armaduras en la velocidad de movimiento. También pueden llevar un poco más de peso.
