+++ SolastaUnfinishedBusiness/Translations/es/Spells/Spells03-es.txt
@@ -12,8 +12,6 @@ Feedback/&AdditionalDamageSpiritShroudDamageRadiantFormat=Sudario de Espíritu!
 Feedback/&AdditionalDamageSpiritShroudDamageRadiantLine=¡{0} inflige {2} daño radiante a {1} de Spirit Shroud!
 Spell/&BlindingSmiteDescription=Tu próximo golpe causa daño radiante adicional con tu arma destellando con una luz brillante. Además, el objetivo debe tener éxito en una tirada de salvación de Constitución o quedará cegado durante 1 minuto.
 Spell/&BlindingSmiteTitle=Golpe cegador
-Spell/&EarthTremorDescription=Golpeas el suelo y desatas un temblor de fuerza sísmica, arrojando tierra, rocas y arena.
-Spell/&EarthTremorTitle=Terremoto
 Spell/&SpiritShroudDamageColdDescription=Inflige daño adicional por frío.
 Spell/&SpiritShroudDamageColdTitle=Embozo de espíritu (frío)
 Spell/&SpiritShroudDamageNecroticDescription=Inflige daño necrótico adicional.
