+++ SolastaUnfinishedBusiness/Translations/fr/Spells/Spells01-fr.txt
@@ -20,6 +20,8 @@ Feedback/&AdditionalDamageWrathfulSmiteFormat=Châtiment courroucé!
 Feedback/&AdditionalDamageWrathfulSmiteLine={0} inflige plus de dégâts à {1} par un châtiment courroucé (+{2})
 Spell/&ChromaticOrbDescription=Vous lancez une sphère d'énergie de 4 pouces de diamètre sur une créature que vous pouvez voir à portée. Vous choisissez acide, froid, feu, foudre, poison ou tonnerre pour le type d'orbe que vous créez, puis effectuez une attaque avec un sort à distance contre la cible. Si l'attaque réussit, la créature subit 3d8 dégâts du type que vous avez choisi.
 Spell/&ChromaticOrbTitle=Orbe chromatique
+Spell/&EarthTremorDescription=Vous frappez le sol et déclenchez un tremblement de force sismique, projetant de la terre, de la roche et du sable.
+Spell/&EarthTremorTitle=Tremblement de terre
 Spell/&EnsnaringStrikeDescription=Votre prochain coup inflige des dégâts perforants supplémentaires et votre cible est prise au piège par des lianes.
 Spell/&EnsnaringStrikeTitle=Frappe envoûtante
 Spell/&MuleDescription=Le destinataire de ce sort est capable d'ignorer les effets des charges lourdes ou de l'armure sur la vitesse de déplacement. Ils peuvent également supporter un peu plus de poids.
