+++ SolastaUnfinishedBusiness/Translations/fr/Spells/Spells03-fr.txt
@@ -12,8 +12,6 @@ Feedback/&AdditionalDamageSpiritShroudDamageRadiantFormat=Voile spirituel!
 Feedback/&AdditionalDamageSpiritShroudDamageRadiantLine={0} inflige {2} dégâts radiants à {1} avec Spirit Shroud!
 Spell/&BlindingSmiteDescription=Votre prochain coup cause des dégâts radiants supplémentaires avec votre arme flamboyante avec une lumière vive. De plus, la cible doit réussir un jet de sauvegarde de Constitution ou être aveuglée pendant 1 minute.
 Spell/&BlindingSmiteTitle=Châtiment aveuglant
-Spell/&EarthTremorDescription=Vous frappez le sol et déclenchez un tremblement de force sismique, projetant de la terre, de la roche et du sable.
-Spell/&EarthTremorTitle=Tremblement de terre
 Spell/&SpiritShroudDamageColdDescription=Inflige des dégâts de froid supplémentaires.
 Spell/&SpiritShroudDamageColdTitle=Voile spirituel (froid)
 Spell/&SpiritShroudDamageNecroticDescription=Inflige des dégâts nécrotiques supplémentaires.
