+++ SolastaUnfinishedBusiness/Translations/it/Spells/Spells01-it.txt
@@ -20,6 +20,8 @@ Feedback/&AdditionalDamageWrathfulSmiteFormat=Punizione adirata!
 Feedback/&AdditionalDamageWrathfulSmiteLine={0} infligge più danni a {1} attraverso una punizione adirata (+{2})
 Spell/&ChromaticOrbDescription=Scagli una sfera di energia del diametro di 4 pollici contro una creatura che puoi vedere a gittata. Scegli acido, freddo, fuoco, fulmine, veleno o tuono per il tipo di sfera che crei, quindi esegui un attacco con incantesimo a distanza contro il bersaglio. Se l'attacco va a segno, la creatura subisce 3d8 danni del tipo scelto.
 Spell/&ChromaticOrbTitle=Sfera cromatica
+Spell/&EarthTremorDescription=Colpisci il suolo e scateni una scossa di forza sismica, sollevando terra, roccia e sabbia.
+Spell/&EarthTremorTitle=Tremore Terrestre
 Spell/&EnsnaringStrikeDescription=Il tuo prossimo colpo provoca danni perforanti aggiuntivi e il tuo bersaglio viene intrappolato dai rampicanti.
 Spell/&EnsnaringStrikeTitle=Colpo Intrappolante
 Spell/&MuleDescription=Il destinatario di questo incantesimo è in grado di ignorare gli effetti di carichi pesanti o armature sulla velocità di movimento. Possono anche portare un po' più di peso.
