+++ SolastaUnfinishedBusiness/Translations/it/Spells/Spells03-it.txt
@@ -12,8 +12,6 @@ Feedback/&AdditionalDamageSpiritShroudDamageRadiantFormat=Spirito Sindone!
 Feedback/&AdditionalDamageSpiritShroudDamageRadiantLine={0} infligge {2} danni radianti a {1} da Spirit Shroud!
 Spell/&BlindingSmiteDescription=Il tuo prossimo colpo provoca danni radianti aggiuntivi con la tua arma che lampeggia con una luce brillante. Inoltre, il bersaglio deve superare un tiro salvezza su Costituzione o rimanere accecato per 1 minuto.
 Spell/&BlindingSmiteTitle=Punizione Accecante
-Spell/&EarthTremorDescription=Colpisci il suolo e scateni una scossa di forza sismica, sollevando terra, roccia e sabbia.
-Spell/&EarthTremorTitle=Tremore Terrestre
 Spell/&SpiritShroudDamageColdDescription=Infliggi danni da freddo bonus.
 Spell/&SpiritShroudDamageColdTitle=Velo spirituale (freddo)
 Spell/&SpiritShroudDamageNecroticDescription=Infliggi danni necrotici bonus.
