+++ SolastaUnfinishedBusiness/Translations/pt-BR/Spells/Spells01-pt-BR.txt
@@ -20,6 +20,8 @@ Feedback/&AdditionalDamageWrathfulSmiteFormat=Golpe Colérico!
 Feedback/&AdditionalDamageWrathfulSmiteLine={0} causa mais dano a {1} por meio de um golpe irado (+{2})
 Spell/&ChromaticOrbDescription=Você arremessa uma esfera de energia de 4 polegadas de diâmetro em uma criatura que você possa ver dentro do alcance. Você escolhe ácido, frio, fogo, relâmpago, veneno ou trovão para o tipo de orbe que você cria e, em seguida, faz um ataque de magia à distância contra o alvo. Se o ataque acertar, a criatura sofre 3d8 de dano do tipo que você escolheu.
 Spell/&ChromaticOrbTitle=Esfera Cromática
+Spell/&EarthTremorDescription=Você atinge o solo e desencadeia um tremor de força sísmica, arremessando terra, rocha e areia.
+Spell/&EarthTremorTitle=Tremor de terra
 Spell/&EnsnaringStrikeDescription=Seu próximo golpe causa dano perfurante adicional e seu alvo é enredado por vinhas.
 Spell/&EnsnaringStrikeTitle=Ataque Enganador
 Spell/&MuleDescription=O destinatário deste feitiço é capaz de ignorar os efeitos de cargas pesadas ou armaduras na velocidade de movimento. Eles também podem carregar um pouco mais de peso.
