+++ SolastaUnfinishedBusiness/Translations/pt-BR/Spells/Spells03-pt-BR.txt
@@ -12,8 +12,6 @@ Feedback/&AdditionalDamageSpiritShroudDamageRadiantFormat=Mortalha Espiritual!
 Feedback/&AdditionalDamageSpiritShroudDamageRadiantLine={0} causa {2} de dano radiante a {1} de Spirit Shroud!
 Spell/&BlindingSmiteDescription=Seu próximo golpe causa dano radiante adicional com sua arma brilhando com uma luz brilhante. Além disso, o alvo deve ter sucesso em um teste de resistência de Constituição ou ficará cego por 1 minuto.
 Spell/&BlindingSmiteTitle=Golpe Cegante
-Spell/&EarthTremorDescription=Você atinge o solo e desencadeia um tremor de força sísmica, arremessando terra, rocha e areia.
-Spell/&EarthTremorTitle=Tremor de terra
 Spell/&SpiritShroudDamageColdDescription=Causa dano de frio adicional.
 Spell/&SpiritShroudDamageColdTitle=Mortalha Espiritual (Frio)
 Spell/&SpiritShroudDamageNecroticDescription=Causa dano necrótico adicional.
