+++ SolastaUnfinishedBusiness/Translations/ru/Spells/Spells01-ru.txt
@@ -20,6 +20,8 @@ Feedback/&AdditionalDamageWrathfulSmiteFormat=Гневный удар!
 Feedback/&AdditionalDamageWrathfulSmiteLine={0} наносит больше урона {1} гневным ударом (+{2})
 Spell/&ChromaticOrbDescription=Вы бросаете энергетическую сферу диаметром 4 дюйма в существо, которое видите в пределах досягаемости. Вы выбираете кислоту, холод, огонь, молнию, яд или гром в качестве типа создаваемой вами сферы, а затем совершаете дальнобойную атаку заклинанием против цели. Если атака попадает, существо получает урон 3к8 выбранного вами типа.
 Spell/&ChromaticOrbTitle=Хроматическая сфера
+Spell/&EarthTremorDescription=Вы ударяете по земле и вызываете дрожь сейсмической силы, подбрасывая землю, камни и песок.
+Spell/&EarthTremorTitle=Землетрясение
 Spell/&EnsnaringStrikeDescription=Ваш следующий удар наносит дополнительный колющий урон, и ваша цель попадает в ловушку лиан.
 Spell/&EnsnaringStrikeTitle=Захватывающий удар
 Spell/&MuleDescription=Получатель этого заклинания может игнорировать влияние тяжелых грузов или брони на скорость передвижения. Они также могут нести немного больший вес.
