+++ SolastaUnfinishedBusiness/Translations/ru/Spells/Spells03-ru.txt
@@ -12,8 +12,6 @@ Feedback/&AdditionalDamageSpiritShroudDamageRadiantFormat=Покров Духа!
 Feedback/&AdditionalDamageSpiritShroudDamageRadiantLine={0} наносит {2} урон светом {1} Покровом Духа.
 Spell/&BlindingSmiteDescription=Ваш следующий удар наносит дополнительный урон излучением, когда ваше оружие вспыхивает ярким светом.
 Spell/&BlindingSmiteTitle=Ослепляющая кара
-Spell/&EarthTremorDescription=Вы ударяете по земле и вызываете дрожь сейсмической силы, подбрасывая землю, камни и песок.
-Spell/&EarthTremorTitle=Землетрясение
 Spell/&SpiritShroudDamageColdDescription=Наносит дополнительный урон холодом
 Spell/&SpiritShroudDamageColdTitle=Покров Духа (Холод)
 Spell/&SpiritShroudDamageNecroticDescription=Наносит дополнительный некротический урон.
