+++ SolastaUnfinishedBusiness/Translations/zh-CN/Spells/Spells01-zh-CN.txt
@@ -20,6 +20,8 @@ Feedback/&AdditionalDamageWrathfulSmiteFormat=愤怒之击！
 Feedback/&AdditionalDamageWrathfulSmiteLine={0} 通过愤怒打击对 {1} 造成更多伤害 (+{2})
 Spell/&ChromaticOrbDescription=你向范围内你能看到的一个生物投掷一个直径 4 英寸的能量球。你选择强酸、寒冷、火焰、闪电、毒药或雷霆作为你创造的球体类型，然后对目标进行远程法术攻击。如果攻击命中，该生物将受到你选择类型的 3d8 点伤害。
 Spell/&ChromaticOrbTitle=彩色宝珠
+Spell/&EarthTremorDescription=你猛力打击地面，释放出震荡之力，飞沙走石。
+Spell/&EarthTremorTitle=地动
 Spell/&EnsnaringStrikeDescription=你的下一击会造成额外的穿刺伤害，并且你的目标会被藤蔓缠住。
 Spell/&EnsnaringStrikeTitle=诱捕打击
 Spell/&MuleDescription=此法术的受术者能忽略重物或盔甲对移动速度的影响。他们也能携带略高的重量
