+++ SolastaUnfinishedBusiness/Translations/zh-CN/Spells/Spells03-zh-CN.txt
@@ -6,8 +6,6 @@ Feedback/&AdditionalDamageSpiritShroudDamageNecroticFormat=魂灵环绕!
 Feedback/&AdditionalDamageSpiritShroudDamageNecroticLine={0} 从魂灵环绕对 {1} 造成 {2} 黯蚀伤害！
 Feedback/&AdditionalDamageSpiritShroudDamageRadiantFormat=魂灵环绕!
 Feedback/&AdditionalDamageSpiritShroudDamageRadiantLine={0} 从魂灵环绕对 {1} 造成 {2} 光耀伤害！
-Spell/&EarthTremorDescription=你猛力打击地面，释放出震荡之力，飞沙走石。
-Spell/&EarthTremorTitle=地动
 Spell/&SpiritShroudDamageColdDescription=造成额外寒冷伤害。
 Spell/&SpiritShroudDamageColdTitle=魂灵环绕(寒冷)
 Spell/&SpiritShroudDamageNecroticDescription=造成额外黯蚀伤害。
